[Refactor] 채널필터링 filterChannels 유틸함수로 분리
